third_party/blink/web_tests/external/wpt/css/css-transforms/animation/transform-interpolation-computed-value-expected.txt

This is a testharness.js-based test.
Found 102 tests; 69 PASS, 33 FAIL, 0 TIMEOUT, 0 NOTRUN.
FAIL Interpolation between translateX(0px) and translateX(50px) gives the correct computed value halfway according to computedStyleMap. assert_equals: The value at 50% progress is as expected expected "translateX(25px)" but got "translate(25px, 0px)"
FAIL Interpolation between translateX(0px) and translateX(50px) gives the correct computed value halfway according to computedStyleMap with zoom active. assert_equals: The value at 50% progress is as expected expected "translateX(25px)" but got "translate(25px, 0px)"
PASS Interpolation between translateX(0px) and translateX(50px) gives the correct computed value halfway according to commitStyles.
FAIL Interpolation between translateX(0%) and translateX(50%) gives the correct computed value halfway according to computedStyleMap. assert_equals: The value at 50% progress is as expected expected "translateX(25%)" but got "translate(25%, 0px)"
FAIL Interpolation between translateX(0%) and translateX(50%) gives the correct computed value halfway according to computedStyleMap with zoom active. assert_equals: The value at 50% progress is as expected expected "translateX(25%)" but got "translate(25%, 0px)"
PASS Interpolation between translateX(0%) and translateX(50%) gives the correct computed value halfway according to commitStyles.
PASS Interpolation between translateY(0%) and translateX(50%) gives the correct computed value halfway according to computedStyleMap.
PASS Interpolation between translateY(0%) and translateX(50%) gives the correct computed value halfway according to computedStyleMap with zoom active.
PASS Interpolation between translateY(0%) and translateX(50%) gives the correct computed value halfway according to commitStyles.
PASS Interpolation between translateX(50px) and translateY(50px) gives the correct computed value halfway according to computedStyleMap.
PASS Interpolation between translateX(50px) and translateY(50px) gives the correct computed value halfway according to computedStyleMap with zoom active.
PASS Interpolation between translateX(50px) and translateY(50px) gives the correct computed value halfway according to commitStyles.
PASS Interpolation between translateX(50px) and translateZ(50px) gives the correct computed value halfway according to computedStyleMap.
PASS Interpolation between translateX(50px) and translateZ(50px) gives the correct computed value halfway according to computedStyleMap with zoom active.
PASS Interpolation between translateX(50px) and translateZ(50px) gives the correct computed value halfway according to commitStyles.
PASS Interpolation between translateZ(50px) and translateX(50px) gives the correct computed value halfway according to computedStyleMap.
PASS Interpolation between translateZ(50px) and translateX(50px) gives the correct computed value halfway according to computedStyleMap with zoom active.
PASS Interpolation between translateZ(50px) and translateX(50px) gives the correct computed value halfway according to commitStyles.
FAIL Interpolation between translateZ(-50px) and translateZ(50px) gives the correct computed value halfway according to computedStyleMap. assert_equals: The value at 50% progress is as expected expected "translateZ(0px)" but got "translate3d(0px, 0px, 0px)"
FAIL Interpolation between translateZ(-50px) and translateZ(50px) gives the correct computed value halfway according to computedStyleMap with zoom active. assert_equals: The value at 50% progress is as expected expected "translateZ(0px)" but got "translate3d(0px, 0px, 0px)"
PASS Interpolation between translateZ(-50px) and translateZ(50px) gives the correct computed value halfway according to commitStyles.
FAIL Interpolation between translate(0%) and translate(50%) gives the correct computed value halfway according to computedStyleMap. assert_equals: The value at 50% progress is as expected expected "translate(25%)" but got "translate(25%, 0px)"
FAIL Interpolation between translate(0%) and translate(50%) gives the correct computed value halfway according to computedStyleMap with zoom active. assert_equals: The value at 50% progress is as expected expected "translate(25%)" but got "translate(25%, 0px)"
FAIL Interpolation between translate(0%) and translate(50%) gives the correct computed value halfway according to commitStyles. assert_equals: The value at 50% progress is as expected expected "translate(25%)" but got "translate(25%, 0px)"
PASS Interpolation between translate(50%) and translate(100%, 50%) gives the correct computed value halfway according to computedStyleMap.
PASS Interpolation between translate(50%) and translate(100%, 50%) gives the correct computed value halfway according to computedStyleMap with zoom active.
PASS Interpolation between translate(50%) and translate(100%, 50%) gives the correct computed value halfway according to commitStyles.
PASS Interpolation between translate(0%, 50%) and translate(50%, 100%) gives the correct computed value halfway according to computedStyleMap.
PASS Interpolation between translate(0%, 50%) and translate(50%, 100%) gives the correct computed value halfway according to computedStyleMap with zoom active.
PASS Interpolation between translate(0%, 50%) and translate(50%, 100%) gives the correct computed value halfway according to commitStyles.
PASS Interpolation between translate3d(0,0,-50px) and translateZ(50px) gives the correct computed value halfway according to computedStyleMap.
PASS Interpolation between translate3d(0,0,-50px) and translateZ(50px) gives the correct computed value halfway according to computedStyleMap with zoom active.
PASS Interpolation between translate3d(0,0,-50px) and translateZ(50px) gives the correct computed value halfway according to commitStyles.
FAIL Interpolation between translate(50px, 0px) and translate(100px, 0px) gives the correct computed value halfway according to computedStyleMap. assert_equals: The value at 50% progress is as expected expected "translate(75px)" but got "translate(75px, 0px)"
FAIL Interpolation between translate(50px, 0px) and translate(100px, 0px) gives the correct computed value halfway according to computedStyleMap with zoom active. assert_equals: The value at 50% progress is as expected expected "translate(75px)" but got "translate(75px, 0px)"
FAIL Interpolation between translate(50px, 0px) and translate(100px, 0px) gives the correct computed value halfway according to commitStyles. assert_equals: The value at 50% progress is as expected expected "translate(75px)" but got "translate(75px, 0px)"
PASS Interpolation between rotate(30deg) and rotate(90deg) gives the correct computed value halfway according to computedStyleMap.
PASS Interpolation between rotate(30deg) and rotate(90deg) gives the correct computed value halfway according to computedStyleMap with zoom active.
PASS Interpolation between rotate(30deg) and rotate(90deg) gives the correct computed value halfway according to commitStyles.
FAIL Interpolation between rotateZ(30deg) and rotateZ(90deg) gives the correct computed value halfway according to computedStyleMap. assert_equals: The value at 50% progress is as expected expected "rotateZ(60deg)" but got "rotate(60deg)"
FAIL Interpolation between rotateZ(30deg) and rotateZ(90deg) gives the correct computed value halfway according to computedStyleMap with zoom active. assert_equals: The value at 50% progress is as expected expected "rotateZ(60deg)" but got "rotate(60deg)"
FAIL Interpolation between rotateZ(30deg) and rotateZ(90deg) gives the correct computed value halfway according to commitStyles. assert_equals: The value at 50% progress is as expected expected "rotateZ(60deg)" but got "rotate(60deg)"
FAIL Interpolation between rotate(0deg) and rotateZ(90deg) gives the correct computed value halfway according to computedStyleMap. assert_equals: The value at 50% progress is as expected expected "rotate3d(0, 0, 1, 45deg)" but got "rotate(45deg)"
FAIL Interpolation between rotate(0deg) and rotateZ(90deg) gives the correct computed value halfway according to computedStyleMap with zoom active. assert_equals: The value at 50% progress is as expected expected "rotate3d(0, 0, 1, 45deg)" but got "rotate(45deg)"
FAIL Interpolation between rotate(0deg) and rotateZ(90deg) gives the correct computed value halfway according to commitStyles. assert_equals: The value at 50% progress is as expected expected "rotate3d(0, 0, 1, 45deg)" but got "rotate(45deg)"
FAIL Interpolation between rotateX(0deg) and rotateX(90deg) gives the correct computed value halfway according to computedStyleMap. assert_equals: The value at 50% progress is as expected expected "rotateX(45deg)" but got "rotate3d(1, 0, 0, 45deg)"
FAIL Interpolation between rotateX(0deg) and rotateX(90deg) gives the correct computed value halfway according to computedStyleMap with zoom active. assert_equals: The value at 50% progress is as expected expected "rotateX(45deg)" but got "rotate3d(1, 0, 0, 45deg)"
FAIL Interpolation between rotateX(0deg) and rotateX(90deg) gives the correct computed value halfway according to commitStyles. assert_equals: The value at 50% progress is as expected expected "rotateX(45deg)" but got "rotate3d(1, 0, 0, 45deg)"
PASS Interpolation between rotate(0deg) and rotateX(90deg) gives the correct computed value halfway according to computedStyleMap.
PASS Interpolation between rotate(0deg) and rotateX(90deg) gives the correct computed value halfway according to computedStyleMap with zoom active.
PASS Interpolation between rotate(0deg) and rotateX(90deg) gives the correct computed value halfway according to commitStyles.
FAIL Interpolation between scale(1) and scale(2) gives the correct computed value halfway according to computedStyleMap. assert_equals: The value at 50% progress is as expected expected "scale(1.5)" but got "scale(1.5, 1.5)"
FAIL Interpolation between scale(1) and scale(2) gives the correct computed value halfway according to computedStyleMap with zoom active. assert_equals: The value at 50% progress is as expected expected "scale(1.5)" but got "scale(1.5, 1.5)"
FAIL Interpolation between scale(1) and scale(2) gives the correct computed value halfway according to commitStyles. assert_equals: The value at 50% progress is as expected expected "scale(1.5)" but got "scale(1.5, 1.5)"
PASS Interpolation between scale(1, 3) and scale(2) gives the correct computed value halfway according to computedStyleMap.
PASS Interpolation between scale(1, 3) and scale(2) gives the correct computed value halfway according to computedStyleMap with zoom active.
PASS Interpolation between scale(1, 3) and scale(2) gives the correct computed value halfway according to commitStyles.
FAIL Interpolation between scaleX(1) and scaleX(2) gives the correct computed value halfway according to computedStyleMap. assert_equals: The value at 50% progress is as expected expected "scaleX(1.5)" but got "scale(1.5, 1)"
FAIL Interpolation between scaleX(1) and scaleX(2) gives the correct computed value halfway according to computedStyleMap with zoom active. assert_equals: The value at 50% progress is as expected expected "scaleX(1.5)" but got "scale(1.5, 1)"
PASS Interpolation between scaleX(1) and scaleX(2) gives the correct computed value halfway according to commitStyles.
FAIL Interpolation between scaleY(1) and scaleY(2) gives the correct computed value halfway according to computedStyleMap. assert_equals: The value at 50% progress is as expected expected "scaleY(1.5)" but got "scale(1, 1.5)"
FAIL Interpolation between scaleY(1) and scaleY(2) gives the correct computed value halfway according to computedStyleMap with zoom active. assert_equals: The value at 50% progress is as expected expected "scaleY(1.5)" but got "scale(1, 1.5)"
PASS Interpolation between scaleY(1) and scaleY(2) gives the correct computed value halfway according to commitStyles.
FAIL Interpolation between scaleZ(1) and scaleZ(2) gives the correct computed value halfway according to computedStyleMap. assert_equals: The value at 50% progress is as expected expected "scaleZ(1.5)" but got "scale3d(1, 1, 1.5)"
FAIL Interpolation between scaleZ(1) and scaleZ(2) gives the correct computed value halfway according to computedStyleMap with zoom active. assert_equals: The value at 50% progress is as expected expected "scaleZ(1.5)" but got "scale3d(1, 1, 1.5)"
PASS Interpolation between scaleZ(1) and scaleZ(2) gives the correct computed value halfway according to commitStyles.
FAIL Interpolation between scaleX(2) and scaleY(2) gives the correct computed value halfway according to computedStyleMap. assert_equals: The value at 50% progress is as expected expected "scale(1.5)" but got "scale(1.5, 1.5)"
FAIL Interpolation between scaleX(2) and scaleY(2) gives the correct computed value halfway according to computedStyleMap with zoom active. assert_equals: The value at 50% progress is as expected expected "scale(1.5)" but got "scale(1.5, 1.5)"
FAIL Interpolation between scaleX(2) and scaleY(2) gives the correct computed value halfway according to commitStyles. assert_equals: The value at 50% progress is as expected expected "scale(1.5)" but got "scale(1.5, 1.5)"
PASS Interpolation between scaleX(2) and scaleY(3) gives the correct computed value halfway according to computedStyleMap.
PASS Interpolation between scaleX(2) and scaleY(3) gives the correct computed value halfway according to computedStyleMap with zoom active.
PASS Interpolation between scaleX(2) and scaleY(3) gives the correct computed value halfway according to commitStyles.
PASS Interpolation between scaleZ(1) and scale(2) gives the correct computed value halfway according to computedStyleMap.
PASS Interpolation between scaleZ(1) and scale(2) gives the correct computed value halfway according to computedStyleMap with zoom active.
PASS Interpolation between scaleZ(1) and scale(2) gives the correct computed value halfway according to commitStyles.
PASS Interpolation between scale(1, 2) and scale(3, 4) gives the correct computed value halfway according to computedStyleMap.
PASS Interpolation between scale(1, 2) and scale(3, 4) gives the correct computed value halfway according to computedStyleMap with zoom active.
PASS Interpolation between scale(1, 2) and scale(3, 4) gives the correct computed value halfway according to commitStyles.
PASS Interpolation between scale3d(1, 2, 3) and scale3d(4, 5, 6) gives the correct computed value halfway according to computedStyleMap.
PASS Interpolation between scale3d(1, 2, 3) and scale3d(4, 5, 6) gives the correct computed value halfway according to computedStyleMap with zoom active.
PASS Interpolation between scale3d(1, 2, 3) and scale3d(4, 5, 6) gives the correct computed value halfway according to commitStyles.
PASS Interpolation between scale3d(1, 2, 3) and scale(4, 5) gives the correct computed value halfway according to computedStyleMap.
PASS Interpolation between scale3d(1, 2, 3) and scale(4, 5) gives the correct computed value halfway according to computedStyleMap with zoom active.
PASS Interpolation between scale3d(1, 2, 3) and scale(4, 5) gives the correct computed value halfway according to commitStyles.
PASS Interpolation between scale(1, 2) and scale3d(3, 4, 5) gives the correct computed value halfway according to computedStyleMap.
PASS Interpolation between scale(1, 2) and scale3d(3, 4, 5) gives the correct computed value halfway according to computedStyleMap with zoom active.
PASS Interpolation between scale(1, 2) and scale3d(3, 4, 5) gives the correct computed value halfway according to commitStyles.
PASS Interpolation between skewX(0deg) and skewX(60deg) gives the correct computed value halfway according to computedStyleMap.
PASS Interpolation between skewX(0deg) and skewX(60deg) gives the correct computed value halfway according to computedStyleMap with zoom active.
PASS Interpolation between skewX(0deg) and skewX(60deg) gives the correct computed value halfway according to commitStyles.
PASS Interpolation between skewX(0deg) and skewX(90deg) gives the correct computed value halfway according to computedStyleMap.
PASS Interpolation between skewX(0deg) and skewX(90deg) gives the correct computed value halfway according to computedStyleMap with zoom active.
PASS Interpolation between skewX(0deg) and skewX(90deg) gives the correct computed value halfway according to commitStyles.
PASS Interpolation between skewX(0deg) and skewX(180deg) gives the correct computed value halfway according to computedStyleMap.
PASS Interpolation between skewX(0deg) and skewX(180deg) gives the correct computed value halfway according to computedStyleMap with zoom active.
PASS Interpolation between skewX(0deg) and skewX(180deg) gives the correct computed value halfway according to commitStyles.
PASS Interpolation between skew(0deg, 0deg) and skew(60deg, 60deg) gives the correct computed value halfway according to computedStyleMap.
PASS Interpolation between skew(0deg, 0deg) and skew(60deg, 60deg) gives the correct computed value halfway according to computedStyleMap with zoom active.
PASS Interpolation between skew(0deg, 0deg) and skew(60deg, 60deg) gives the correct computed value halfway according to commitStyles.
PASS Interpolation between skew(45deg, 0deg) and skew(0deg, 45deg) gives the correct computed value halfway according to computedStyleMap.
PASS Interpolation between skew(45deg, 0deg) and skew(0deg, 45deg) gives the correct computed value halfway according to computedStyleMap with zoom active.
PASS Interpolation between skew(45deg, 0deg) and skew(0deg, 45deg) gives the correct computed value halfway according to commitStyles.
Harness: the test ran to completion.
